% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMVAQ.R
\name{scoreMVAQ}
\alias{scoreMVAQ}
\title{Score the Michigan Vision-related Anxiety Questionnaire}
\usage{
scoreMVAQ(dat, verbose = 0)
}
\arguments{
\item{dat}{data.frame with N rows and 15 columns.  First column is patient ID and the remaining columns are responses to the 14 items.  Responses should be integers from 0 to 4.}

\item{verbose}{A number indicating the amount of printing to console during function execution. 0 (default) is none. Higher numbers may result in more printing.}
}
\value{
A list with two components, each a data.frame with 3 variables: ID, ARF, and ACF. The first component, named thetas, contains the IDs (column 1 of of the input), and the 2 domain scores (ARF, ACF) for each patient.  The second, named ses, contains the IDs and the standard errors of the thetas.
}
\description{
Score the Michigan Vision-related Anxiety Questionnaire
}
\examples{
justme <- read.table(
  text = "Me,0,0,0,0,0,0,0,0,1,1,1,0,0,1",
  sep = ",")
scoreMVAQ(justme)

justyou <- data.frame(ID="You", matrix(nrow=1, sample(0:3, 14, TRUE)))
scoreMVAQ(justyou)

names(justme) <- names(justyou) <- c("ID", sprintf("Q\%0.2d", seq(14)))
justus <- rbind(justme, justyou)
scoreMVAQ(justus, verbose = 1)

}
